requirements-dev.txt
black
uuid
faker
kafka-python
python-decouple
psycopg2-binary
sqlalchemy
argparse
